muitos ajustes e atualizações de movimentações

+++ Motorcycle/FrontEnd/src/pages/CadastroProduto.jsx
@@ -14,7 +14,7 @@ const CadastroProduto = () => {
   const [categorias, setCategorias] = useState([]);
 
 
-   useEffect(() => {
+  useEffect(() => {
     buscarProdutos();
   }, []);
 
@@ -26,7 +26,7 @@ const CadastroProduto = () => {
       const res = await fetch("http://localhost:3000/api/produto");
       const data = await res.json();
 
-       if (!res.ok) {
+      if (!res.ok) {
         alert(data.message);
         return;
       }
@@ -42,21 +42,26 @@ const CadastroProduto = () => {
   }
 
   const deletarProduto = async (id) => {
+    const confirmDelete = window.confirm(
+      "Deseja realmente excluir esta movimentação?"
+    );
+
+    if (!confirmDelete) return;
     setLoading(true)
 
-     try {
+    try {
       const res = await fetch(`http://localhost:3000/api/produto/${id}`, {
         method: 'DELETE'
       });
       const data = await res.json();
 
-       if (!res.ok) {
+      if (!res.ok) {
         alert(data.message);
         return;
       }
-        alert(data.message);
+      alert(data.message);
 
-        buscarProdutos();
+      buscarProdutos();
 
     } catch (error) {
       console.log("erro ao carregar dados:", error)
@@ -70,68 +75,69 @@ const CadastroProduto = () => {
   return (
     <div className="container">
 
-    <div className="topo">
-      <button
-        onClick={() => {
-          setModo("cadastrar");
-          setProduto(null);
-          setOpen(true);
-        }}
-      >
-        + Cadastrar Produto
-      </button>
-    </div>
-
-    {loading ? (
-      <p>Carregando...</p>
-    ) : (
-      <div className="cards">
-        {produtos.map((item) => (
-          <div className="card" key={item.id}>
-
-            <img
-              src={item.image}
-              alt={item.nome}
-              className="imagem"
-            />
-
-            <h2>{item.nome}</h2>
-
-            <p><strong>Marca:</strong> {item.marca}</p>
-            <p><strong>Peso:</strong> {item.peso}</p>
-            <p><strong>Preço:</strong> R$ {item.preco}</p>
-
-            <p>
-              <strong>Categoria:</strong>{" "}
-              {item.Categorium?.nome}
-            </p>
-
-            <div className="acoes">
-              <button
-                onClick={() => {
-                  setModo("editar");
-                  setProduto(item);
-                  setOpen(true);
-                }}
-              >
-                Editar
-              </button>
-
-              <button
-               onClick={() => deletarProduto(item.id)}
-              >
-                Deletar
-              </button>
-            </div>
-
-          </div>
-        ))}
+      <div className="topo">
+        <button
+          onClick={() => {
+            setModo("cadastrar");
+            setProduto(null);
+            setOpen(true);
+          }}
+        >
+          + Cadastrar Produto
+        </button>
       </div>
-    )}
 
-    <ModalProduto open={open} onClose={() => setOpen(false)} produto={produto} modo={modo} onProdutoCadastrado={buscarProdutos}/>
+      {loading ? (
+        <p>Carregando...</p>
+      ) : (
+        <div className="cards">
+          {produtos.map((item) => (
+            <div className="card" key={item.id}>
 
-  </div>
+              <img
+                src={item.image}
+                alt={item.nome}
+                className="imagem"
+              />
+
+              <h2>{item.nome}</h2>
+
+              <p><strong>Marca:</strong> {item.marca}</p>
+              <p><strong>Peso:</strong> {item.peso}</p>
+              <p><strong>Preço:</strong> R$ {item.preco}</p>
+              <p><strong>Estoque:</strong> {item.estoque}</p>
+
+              <p>
+                <strong>Categoria:</strong>{" "}
+                {item.Categorium?.nome}
+              </p>
+
+              <div className="acoes">
+                <button
+                  onClick={() => {
+                    setModo("editar");
+                    setProduto(item);
+                    setOpen(true);
+                  }}
+                >
+                  Editar
+                </button>
+
+                <button
+                  onClick={() => deletarProduto(item.id)}
+                >
+                  Deletar
+                </button>
+              </div>
+
+            </div>
+          ))}
+        </div>
+      )}
+
+      <ModalProduto open={open} onClose={() => setOpen(false)} produto={produto} modo={modo} onProdutoCadastrado={buscarProdutos} />
+
+    </div>
   )
 }
